chore: Update readme examples
In order to ensure better accuracy in the readme usage examples, this
commit fixes some formatting issues and removes the # characters as
these were left over from when the examples lived in lib.rs as rustdoc
directly. The # allowed the code to be used for compilation but not
shown in the actual examples.