Chore: Update README.md examples to be checked for compile-ability
In order to ensure any examples written in the README.md and included
into the lib.rs are checked that they actually compile correctly, this
commit replaces the ignore directive on the code fence for no_run
instead. This ensures during `cargo test` runs that these code blocks
are checked that they are compile-able as they are considered doc-tests.
But it also makes sure we don't actually try and run the code itself as
we don't want to hit real SendWithUs api's during test runs.